Add ComicListView and rework comic loading system
The comic loading system was switched from a completion callback
system to an async loading system which both avoids callback hell and
allows for code segmentation.  The latter was necessary when getting
the latest comic number for the list view.  With the new system it is
possible to get just the metadata or just the image.  It will also
facilitate the addition of a comic cache in the future.