Add persistant image cache with Core Data
I'm too tired to write a full commit message, but the important thing
to note is ComicMetadata.comicImage is optional but
ComicImage.comicMetadata is not.  This must be taken into
consideration when creating ComicImages.  It is an error to have a
ComicImage with no ComicMetadata in the model (This isn't enforced
and afaik there's no way to perfectly enforce this).