Make PersistenceProvider threadsafe and add caches to ComicStore
When I first started using Core Data to store comic data, I didn't
know about the threading restrictions on NSManagedObjectContexts and
NSManagedObjects.  I was creating, fetching, and passing the latter
around between two different DispatchQueues which was causing
crashes.  The solution to this is to perform all Core Data related
operations on the context's DispatchQueue by surrounding the
offending code in context.perform(block:).  Additionally, outside of
PersistenceProvider, comic metadata is now stored in a threadsafe
immutable struct called SafeComicMetadata.  This prevents
ComicMetadata objects being passed to threads differnt from their
context's.  Credit to
https://agiokas.medium.com/core-data-and-async-await-thread-safe-f96b6dbbb7c4
for the ToSafeType protocol and pattern.

Next, after the data was threadsafe, I noticed image loading from
persistance was slower than it used to be.  I'm assuming it was
caused by all the loading tasks needing to wait in the context's
dispatch queue.  The solution to this was adding two NSCaches in
the ComicStore to cache image and metadata structs respectively.
This allows previously loaded images to appear much snappier.