Add UIPageView for swipeable fullscreen comics
The first thing I tried to achieve the goal of swipable comic pages was
to use a TabView with the .page tabViewStyle, however the long press
gesture to view the current comic's alt text was interfering with the
TabView's swip gesture to move between comics.  The current, tentative
solution is to embed a UIPageViewController in a SwiftUI
UIViewControllerRepresentable view.  This, in turn, embeds a
ComicImageViewLongPressWrapper view (which adds the long press action to
the ComicImageView it wraps) inside a UIHostingController to display
comics.  This UIKit sandwhich nearly yields the desired behaviour,
however the user must release the long press before it registers.  This
will likely be solved with some custom UIGestureRecognizers that don't
conflict.  However, currently this is the closest I've come to the
desired behaviour in numerous experiments.

Addtionally a custom modifier has been added which shows a border around
a view's frame and displays it's size in points as an overlay.  This is
very useful for debugging view sizes and other aspects of SwiftUI.  It
is not intended to be shipped in production builds.