Fix long press and UIPageView conflict
After all the trials and errors long pressing a comic to show the alt
text finally works alongside swiping between comics in the
UIPageViewController.  The way this was finally resolved was moving the
long press recognition from SwiftUI's .onLongPress modifier to a
UILongPressGestureRecognizer in the UIHostingController for comics in
fullscreen mode (the recognizer is attached to the hosting controller's
UIKit-land view).  The gesture recognizer's action is a function in the
hosting controller that calls a show alt text function in the hosted
SwiftUI view.  This function finally toggles the alt being shown.
Because the toggle happens outside of the normal SwiftUI callstack and
outside of body, the isAltShown boolean couldn't be a @State variable or
wrapped in a @StateObject.  Instead it had to be wrapped in an
@ObservedObject variable.

This solution feels a little janky and involved, but after so much time
spent stuck on this problem, I'm more than ok accepting it as the
defacto solution.  In all my searching online I couldn't find any
reference to this issue nor any solutions to it.  There may be some
semantic or technical details about SwiftUI gestures I'm missing, but
in all my research I couldn't find anything deeper than surface level
on the matter.  It seems to be hard to find advanced details on SwifUI
and it's components / workings.

I'm not sure how to share this information with anyone else struggling
to rectify this issue, but maybe there's a chance they'll stumble upon
this commit message and get the help they need.