Add custom waterfall UICollectionViewLayout
This took way more effort and debugging than anticipated.  However I'm
really proud of the result.  Right now it doesn't look like the layout
is a waterfall, but enabling Debug->View Debugging->Show View Frames
shows that the waterfall layout is working correctly.  The reason it
looks so janky right now is that the layout isn't adapting the cell
heights to the adjusted height of the image (based on column width and
image aspect ratio).  Instead, right now, it is generating a list of
random height for the cells.  It's consistend within a single run of
the app, but changes on every launch.  This is just for testing and
will definitely be changed later.

The biggest hurdle I had to overcome with the layout was an issue
getting the UICollectionView's bounds rect.  Since I'm rogrammatically
creating a UICollectionViewController (as opposed to loading it from a
storyboard or nib file), both the controller's wrapper view and the
collection view had bounds rects of 0.0x0.0.  In all the code references
I found online (including multiple from Apple), the first data snapshot
(of the diffable data source) was applied in viewDidLoad().  However,
when I was doing the same I was running into the 0x0 bounds issue.  This
issue was chained into my layout controller as the collection view's
layout's prepare() function is called from
UICollectionViewDiffableDataSource.apply(snapshot:).  As the width of
the collection view cells is calculated based on the width of the
collection view and the number of columns, a width of 0 was causing
major issues.  The final fix for this was moving the first snapshot
application from viewDidLoad() to viewWillAppear().  By then the
collection view has proper bounds assigned to it.  The assumption I have
come to is that auto layout happens between the two lifecycle functions
of the controller.  Ultimately, if a UICollectionViewController is
created programmatically, its view and collection view won't have bounds
before at least viewWillAppear() (and therefore a diffable data source
can't be applied until then (if the layout relies on the dimensions of
the collection view)).

Another weird issue I ran into was when embedding the
UIViewControllerRepresentable view into a swiftui navigation view, the
LargeTitleNavigationBar (might not be exactly what it was called) was
interfering with the collection view.  It would display the initial set
of comics, but when any scrolling happened it would go black and
wouldn't recover.  It seems to have mysteriously resolved itself and I
don't remember if I changed anything to fix it.  Just something to look
out for in the future.

Speaking of the future, currently the waterfall layout is using the slow
way of implementing layoutAttributesForElements(in:).  It is iterating
over all attributes in the cache, however there is a much more efficient way
to fufill the request with a binary search.  The binary search is
written, but not currently working.  It only shows a select few comics,
then nothing after.  The slow method seems to not impede performance,
but it's not a long term solution so it'll be fixed later.

Also also, I noticed some hitching / abrupt abrupt resizing of the
cells when the collection view first loads.  Toggling view borders shows
that the cells themselves aren't resizing so I'm going to chalk it up to
SwiftUI shenanigans.  Hopefully, the new SwiftUI views I make to
implement for cells fixes the issue.  If not, it will be fixed after
further investigation.

That's it.
...
FIN