Rework model/storage layer and add image ratio
Legacy code is always the worst code.  Especially if you're the working
on a project alone.  I went through and streamlined a lot about the
storage system.

The biggest change I made was to ComicMetadata (formerly
SafeComicMetadata).  It is now Decodable and replaces both
SafeComicMetadata and JsonComicMetadata.  It is what is both decoded
from the json and passed around in a thread-safely.  This percipitated
the change to Persistence Provider's save functions which no longer
return the thread safe version of the object they are storing.

Additionally, I removed the code duplication in
getLatestStoredMetadata/Blocking() and now just wrap the synchronous
function in a task for the async version.  Not sure if this truly makes
the call asynchronous but I think it helps.

Finally, I added image aspect ratio (height / width) storage and
calculation in preperation for cell resizing in
UIComicCollectionViewLayout.  It is both stored in the ComicStore cache
and the Persistence Provider's storage.  That's about it.  Thanks for
watching.