Add ratio based heights to UIComicCollectionViewLayout
When laying out its collection view, UIComicCollectionViewLayout now
sets the cell height to itemWidth * imageRatio (height_px / width_px) if
the comic store has a ratio stored.  Otherwise the cell height is set to
cell width.  This had the very unexpected side-effect of improperly
sizing the hosted SwiftUI view.  The aspect ratio of the view is
correct, but its bounds are inset by an nuknown factor.  Weirder is the
fact that it doesn't happen to all cells and appears to be linked to
scroll speed.  Weirdest of all is this problem is fixed by tapping on
the cell.  No gesture recognizers are installed that should exhibit this
behaviour and yet the view updates its bounds anyway.  Additionally, the
issue seems to resolve itself when the cell scrolls off screen, and
never occurs for the same comic twice.

Because of this inexplicable and seemingly undebuggable error, the cells
will be rewritten in UIKit in the future.

Additonally, some of the fetch methods in PersistenceProvider have been
optimized, along with a new function to retrieve all currently stored
ratios.  This list is then stashed in the layout object.  The single
fetch request makes initial loading many times faster, and the caching
makes all subsequent loads instant (at least in appearance).

NOTE: This cache must be updated as comic images are downloaded.