Start converting cell contents to UIKit
Because of the weird issues embedding SwiftUI views in collection view
cells with my custom waterfall layout, I've began the process of
converting the cell's contents to UIKit views.  The comic image and
number display correctly (the number label needs more work for the
background and font), but the performance still isn't where it should
be.

To pivot (both out of frustration and a hope that it will alleviate
performance issues) I'm going to rewrite the entire model layer.  This
was one of the first components I wrote for this app and it's in dire
need of love and care.  I've learned a ton about swift and iOS dev since
then which I hope to apply to a brand spanking new system with a much
cleaner api.  Additionally, I'm going to move from storing images in the
database to the filesystem since it's way too much binary data to stick
in there (especially with 2x images and storing both lods).

Hopefully it shouldn't be too bad :)  I'll see you on the other side.