Add new ComicFetcher
This is the first component of the new model layer for the app.  It
vastly streamlines the logic into 3 functions (2 of which are public)
and uses tasks and a swift actor to ensure there is only ever one
request to the server at a time for a particular file (metadata or comic
image of either size).

The app doesn't currently compile because the rest of the model layer
is missing (notably the ComicStore which was and will continue to be
the public interface).

The next step is to save the comic files to the filesystem and create a
pseudo database api on top of it.