Add ComicImageDatabase
Added an api wrapper around saving images to and loading images from
disk.  This was tested in a test project on a mac, but on on device.
The app still doesn't compile as ComicStore hasn't bene reimplemented.
ComicImageDatabase throws errors instead of returning optionals like
older implementations of the model layer.

Additionally, ComicFetcher was updated to throw errors instead of
returning optionals.  Not only is this more idiomatic code and
eliminates returned optionals, but it also conveys information about the
cause of the error to the calling scope.  This change hasn't been
tested, but it shouldn't have affected any functionality.