Fill out PersistenceProvider
With PersistenceProvider done, the only thing left to do for the model
layer is to make a ComicStore public interface.  I tried to design the
api to be more generalized, but may end up needing to add ad-hoc
features in the future.

This section hasn't been tested before being committed so it may need
debugging later.