Add and integrate new ComicStore api
I finally made a new ComicStore class to unify the api contract into the
model layer.  It is much cleaner than the last one, but lacks the last
necessary piece: saving image ratios.  That is the next and (hopefully)
final step the model layer needs to have feature pairity with the old
one.

There are two major things to note here.  First, is that when scrolling
too far down in the collection view, images stop loading and therefore
cells remain blank.  This seems to be an issue relating to the new
single task nature of comic fetching / loading.  Further investigation
is required.

The second notable thing is that, even with the new model layer, there
is still a stutter while scrolling in the collection view.  I discovered
in a wwdc 2021 talk ("Make blazing fast lists and collection views")
that this is being caused by image decompression on the main thread when
setting a UIImageView's image property.  A UIImage can either be in a
compressed or uncompressed (bitmap) state.  Uncompressed images are
required for rendering so it is uncompressed when first setting the
image, blocking the main thread and causing scrolling hitches.  I've
begun further exploring solutions to this and will include them in a
future commit, but wanted to isolate this commit to the new ComicStore.
The reason I'm noting this here is to clarity that the hitching isn't
the fault of the new model layer.  It appears to be fairly solid and
future-proof.  Hopefully it won't need another drastic reformation.