Fix scroll stutters in collection view
It has been a long path to this fix.  First, I rewrote the model layer.
That didn't fix it.  Next, after integrating the new model layer in the
existing code, images before a certian point wouldn't display.  The new
model layer wasn't falling back to default sized images when a large
image didn't exist.  Then, having found a solution to the stuttering
in a wwdc 2021 talk (Make blazing fast lists and collection views), I
tried implementing asynchronous image preperation.  That failed on a
specific (and seemingly random) range of images (as well as spitting out
an error on every image).  I wrote a png chunk dumper to view the
differences between sucessful and failing images (a readily available
one didn't exist).  This finally lead me to the conclusion
UIImage.byPreparingForDisplay() fails on pngs with palette/indexed
colors (even though displaying an image without this asynchronous
processing worked on all images).  Finally, the solution that finally
worked ended up being a workaround.  Instead of calling
UIImage.bypreparingForDisplay(), I called
UIImage.byPreparingThumbnail(ofSize:) and passed the image's size in
pixels (CGSize(width: image.size.width * image.scale, height:
image.size.width * image.scale)).

The are a couple things along the way to make a note of.  First, there
is a new utility view for displaying the contents of a directory in a
hierarchical list called DirectoryExplorerView.  Eventually, I intend
to spin off all the utilities I build up for this project into a
seperate utilities library (show frame bounds modifier, directory
explorer, etc.).

Second, is the update to ComicImageDatabase's url handling.  Originally,
I was running into walls with constructing urls relativeTo other urls.
This was failing in weird and inexplicable ways when trying to use them
for file io.  Because of this, I switched the url construction to string
manipulation and just made the final url as needed.  This time around,
I learned about the appendingPathComponent() method of urls which
return a new url with the component appended.  This is the method of
constructing urls from now on.  (This probably isn't notable for most
people, but I ran into so many frustrations that I wanted to document
it.)