Refactor code and remove unused test code
This also pins the realm-swift package at version 10.41.0. It was
previously set (in the package dependencies tab of the xkcdz project
file) to "Up to Next Major Version", however in realm-swift 10.49.3
the linking process was changed (from linking both the Realm _and_
RealmSwift packages to just Realm _or_ RealmSwift) which causes issues
automatically (without user interaction).

This can (and probably) should be bumped up to the 10.49.3 and set back
to "Up to Next Major Version", but for now, this is the quicker and
easier solution.